requirements.txt
Django==5.2.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0